requirements.txt
numpy==1.24.3
matplotlib==3.7.1
pillow==10.0.0
scikit-learn==1.3.0
opencv-python==4.8.0.76
